+++ src/cpu/x64/jit_gemm_inner_product_utils.cpp
@@ -313,7 +313,7 @@ jit_pp_kernel_t<isa>::jit_pp_kernel_t(size_t OC, size_t MB, dim_t dst_mb_stride,
 #define PARAM_OFF(field) offsetof(ker_args_t, field)
         static constexpr bool preserve_gpr = true;
         static constexpr bool preserve_vmm = true;
-        static const size_t helper_vmm_idx = is_avx512_ ? 31 : 15;
+        static const int helper_vmm_idx = is_avx512_ ? 31 : 15;
         static constexpr bool use_exact_tail_scalar_bcast = false;
         const auto dst_md_wrapper = memory_desc_wrapper(*dst_md);
 
@@ -326,7 +326,7 @@ jit_pp_kernel_t<isa>::jit_pp_kernel_t(size_t OC, size_t MB, dim_t dst_mb_stride,
             OC_loop = vlen * max_OC_loop_unroll_;
             OC_tail = OC % OC_loop;
         }
-        size_t tail_size = OC_tail % vlen;
+        int tail_size = static_cast<int>(OC_tail % vlen);
         // enable tail processing for runtime load even if there is no tail
         // for the OC
         tail_size = !!tail_size ? tail_size : 1;
